pyproject.toml
[project]
name = "a8wisper"
version = "0.1.0"
description = "A8Wisper - Windows AI Voice Assistant (Hybrid UI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "llama-cpp-python>=0.2.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "keyboard>=0.13.5",
    "huggingface-hub>=0.20.0",
    "pyautogui>=0.9.54",
    "openai>=1.0.0",
    "pyperclip>=1.9.0",
    "pywebview>=5.0",
    "pystray>=0.19",
    "pillow>=10.0.0",
    "bottle>=0.12.0", # Reliable internal server for pywebview
    "websocket-client>=1.6.0",
    "PySide6>=6.5.0",  # Qt6 for improved WebView window
    "websockets>=12.0", # Async WebSocket server
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch" }]
torchvision = [{ index = "pytorch" }]
torchaudio = [{ index = "pytorch" }]

[dependency-groups]
test = [
    "hypothesis>=6.148.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
